completed stopwatch
