+++ Frontend/src/Styles/Calendar.css
@@ -567,7 +567,7 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    padding: 20px;
+    padding: 5px;
 }
 
 .add-event-footer .add-event-btn {
